+++ browserup-proxy-core/src/main/java/com/browserup/harreader/model/HarHeader.java
@@ -1,9 +1,0 @@
-package com.browserup.harreader.model;
-
-/**
- * @deprecated Use {@link de.sstoehr.harreader.model.HarHeader}
- */
-@Deprecated
-public class HarHeader extends de.sstoehr.harreader.model.HarHeader {
-
-}