+++ browserup-proxy-core/src/main/java/com/browserup/harreader/model/HarRequest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import de.sstoehr.harreader.model.*;
 
 import java.util.ArrayList;
 import java.util.HashMap;