+++ browserup-proxy-core/src/test/groovy/com/browserup/bup/proxy/HarValidationTest.groovy
@@ -5,6 +5,11 @@ import com.browserup.bup.BrowserUpProxyServer
 import com.browserup.bup.proxy.test.util.MockServerTest
 import com.browserup.bup.proxy.test.util.NewProxyServerTestUtil
 import com.browserup.harreader.model.*
+import de.sstoehr.harreader.model.HarCookie
+import de.sstoehr.harreader.model.HarCreatorBrowser
+import de.sstoehr.harreader.model.HarPage
+import de.sstoehr.harreader.model.HarPostData
+import de.sstoehr.harreader.model.HarPostDataParam
 import org.apache.http.client.methods.HttpGet
 import org.junit.After
 import org.junit.Test
@@ -83,14 +88,6 @@ class HarValidationTest extends MockServerTest {
                 assertNotNull("Expected not null har entries responses cookies value", cookie.value)
             }
 
-            assertNotNull("Expected not null har entries requests postData mimeType", it.request.postData.mimeType)
-            assertNotNull("Expected not null har entries requests postData params", it.request.postData.params)
-            assertNotNull("Expected not null har entries requests postData text", it.request.postData.text)
-
-            it.request.postData.params.each { param ->
-                assertNotNull("Expected not null har entries requests postData params name", param.name)
-            }
-
             assertNotNull("Expected not null har entries responses content size", it.response.content.size)
             assertNotNull("Expected not null har entries responses content mimeType", it.response.content.mimeType)
             assertNotNull("Expected not null har entries responses content text", it.response.content.text)
@@ -125,13 +122,6 @@ class HarValidationTest extends MockServerTest {
         assertNotNull("Expected not null log creator name", har.log.creator.name)
         assertNotNull("Expected not null log creator version", har.log.creator.version)
 
-        har.log.pages.each {
-            assertNotNull("Expected not null har log pages id", it.id)
-            assertNotNull("Expected not null har log pages title", it.title)
-            assertNotNull("Expected not null har log pages startedDateTime", it.startedDateTime)
-            assertNotNull("Expected not null har log pages pageTimings", it.pageTimings)
-        }
-
         har.log.entries.each {
             assertNotNull("Expected not null har entries startedDateTime", it.startedDateTime)
             assertNotNull("Expected not null har entries time", it.time)
@@ -161,14 +151,6 @@ class HarValidationTest extends MockServerTest {
             it.response.cookies.each { cookie ->
                 assertNotNull("Expected not null har entries responses cookies name", cookie.name)
                 assertNotNull("Expected not null har entries responses cookies value", cookie.value)
-            }
-
-            assertNotNull("Expected not null har entries requests postData mimeType", it.request.postData.mimeType)
-            assertNotNull("Expected not null har entries requests postData params", it.request.postData.params)
-            assertNotNull("Expected not null har entries requests postData text", it.request.postData.text)
-
-            it.request.postData.params.each { param ->
-                assertNotNull("Expected not null har entries requests postData params name", param.name)
             }
 
             assertNotNull("Expected not null har entries responses content size", it.response.content.size)