+++ browserup-proxy-core/src/test/groovy/com/browserup/bup/proxy/NewHarTest.groovy
@@ -1,13 +1,10 @@
 package com.browserup.bup.proxy
 
-import com.browserup.harreader.model.HarHeader
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.google.common.collect.Iterables
 import com.browserup.bup.BrowserUpProxy
 import com.browserup.bup.BrowserUpProxyServer
 import com.browserup.harreader.model.Har
-import com.browserup.harreader.model.HarContent
-import com.browserup.harreader.model.HarCookie
 import com.browserup.harreader.model.HarEntry
 import com.browserup.harreader.model.HarResponse
 import com.browserup.harreader.model.HarTiming
@@ -15,6 +12,9 @@ import com.browserup.bup.filters.util.HarCaptureUtil
 import com.browserup.bup.proxy.dns.AdvancedHostResolver
 import com.browserup.bup.proxy.test.util.MockServerTest
 import com.browserup.bup.proxy.test.util.NewProxyServerTestUtil
+import de.sstoehr.harreader.model.HarContent
+import de.sstoehr.harreader.model.HarCookie
+import de.sstoehr.harreader.model.HarHeader
 import org.apache.http.client.config.RequestConfig
 import org.apache.http.client.methods.CloseableHttpResponse
 import org.apache.http.client.methods.HttpGet